refactor airqloud uptime chart into independent component
